Рефакторинг: 1. Структуры перенесены в модуль api_model.py 2. Код приведен в соответствие к PEP8 (кроме __init__.py) 3. Добавлено описание структур и атрибутов
